Added PhotinoWindow API class
To be more flexible down the line when we want to add more
functionality, we need to make sure to name our PhotinoX
classes in a way that prevents name conflicts and makes
clear what the intend behind the class is.

The PhotinoNET class exports only Window related functionality
like creating, moving and resizing and registering handlers and
events. Thus the name change to PhotinoWindow.