Make use of Invoke method
The API started to behave in unexpected ways when using more than one
window at a time. Sending a web message into a window view would crash
the application for example.

Using the Invoke method to ensure certain processes are run on the
correct thread is bringing more stability to the API. Not everything
is working though. One problem that persists is that an alert window
opened with OpenAlertWindow() crashes the application when it is
closed.

Another problem that might be unrelated is that the application
crashes when a navigation (using Load) occurs after the instance's
WaitForClose() method was fired. More investigation is needed.