Remove auto-show from constructor
The Show() method of the PhotinoWindow instance
has to be called before a navigation with Load() can
happen. Until now the Show() method was called in
the constructor of the class. This means that the
window is visible before all initial Resize() and MoveTo()
commands are processed. This leads to rendering issues.

Since the window has to be shown before a navigation
can happen, we check for whether the window was
previously shown before the native navigation methods
are called.