Check the AppContext BaseDir for files
When running an application as a single-file published
app bundle, under certain circumstances (like in macOS),
files are not found because the application doesn't look
in the correct directory in which the resource files are
saved.

To prevent this, check whether the file exists under the
full path determined the old way. If not found, check
the app context base directory for the file as well.
If there is no matching file found still, write some error
output to console for debugging purposes. This part must
still be refined with a better way of notifying developers
about the missing file.